Adds cell highlighting feature
Cells can now be highlighted bright red. Highlighting mode is toggled by pressing H. When highlighting mode is activated, the currently selected cell(s) are highlighted. If a single cell is highlighted, the currently highlighted cell can be selected with the + and - buttons, which cycles which cell is currently being highlighted. The highlighted cell(s) along with whether highlighting is enabled is displayed in text labels in the top right of the screen. Up to two cells can be selected at a time. Also, clicking an empty region of the frame now has the equivalent effect as pressing escape and deselects any selected cells.