Fixes highlighting bug
In previous commit, it was possible to highlight a cell ID that did not actually exist in the file. This problem was caused by the num_tracks variable was incorrectly incremented upon the file being opened. This commit also fixes a glitch with the contrast adjustment feature, which also relied on the num_tracks variable, which previously permitted the contrast to be adjusted too far which would cause all the cells to be rendered black.