add flexibility in opening and saving npz files
Previously, npz mode would try to open "raw" and "annotated" files from
provided npz. However, many of our npzs are saved as "X" and "y". Explicitly
added an attempt to open "X" and "y" from npz and remember what files were named.
If data was opened from "raw" and "annotated", it is saved with those filenames.
Otherwise, it is saved to "X" and "y" variable names in the new npz.