change key looked for when creating self.display_info in npz mode
If there is no "cell 1" in the npz (possible especially when working with
pieces of much larger npzs), initialization of npz mode would fail with a
KeyError. Instead of looking through self.cell_info for the keys in feature 0
cell 1 info dict, finds the first key in that dictionary and uses the info keys
found within that cell's info dictionary.