use helper functions to add or delete cell info from dict
Each function (in npz mode) that adds or removes a cell from a frame
or whole npz now uses helper functions to do so. This streamlines
the code and reduces chances of one add/del piece working incorrectly.

Also, making a new label from keybind commands (eg, "new cell stack")
is now based on the max label in the npz and not num_cells, as using
edit_mode to add in annotations may result in mislabeling based on num_cells.