add save_as_trk option to npz mode
Save_as_trk will create a lineage and save the current channel
and feature together as a .trk file. The filename will be the name
of the npz, plus "_c{}_f{}" to indicate which channel and feature
were used. The lineage file will have frame and label information
but no parent/daughter information.

Option to save as .trk will only be displayed if an npz is open.
mode.py will be passed the argument "filetype = npz" to toggle
the change in display behavior.