add "pick color" tool to edit mode (trk and npz editors)
While in edit mode, use "p" (for "pick") to switch to a
mode prompt. Then click on a non-zero value to change the edit_value
(the value the brush can draw or erase) to the selected value. Should
make annotation fixes faster by removing need to scroll through values,
especially in files with many annotations.