add option to sequentially relabel cells in frame of npz
Sequentially relabels cells in the current frame and feature
of npz, starting from 1. Can be useful when correcting sub-images
taken from larger image set or movie, when annotations don't
necessarily start from 1. Not added to trk mode because this only changes
annotations in one frame, which would mess with sequential frames in
a trk.