update predict_zstacks to add only non-zero values
Slight change to make first relabeling step more clear. In first
pass of matching cells between frames, relabeled_image would
get "updated" with zeros if the background was the closest match
to a cell. Now, relabeled_image is only updated with non-zero values
and the step is explicitly skipped when the closest match is background.