Feature/trk feature updates (#36)
* fix typo

one line has sel. instead of self.

* add highlight feature to npz mode

-does not apply in edit mode
-highlight behavior is not completely the same as in .trk version
-does not include "click on background to deselect" behavior
-does not include "click on a third cell to reselect" behavior

* add flexibility in opening and saving npz files

Previously, npz mode would try to open "raw" and "annotated" files from
provided npz. However, many of our npzs are saved as "X" and "y". Explicitly
added an attempt to open "X" and "y" from npz and remember what files were named.
If data was opened from "raw" and "annotated", it is saved with those filenames.
Otherwise, it is saved to "X" and "y" variable names in the new npz.

* change key looked for when creating self.display_info in npz mode

If there is no "cell 1" in the npz (possible especially when working with
pieces of much larger npzs), initialization of npz mode would fail with a
KeyError. Instead of looking through self.cell_info for the keys in feature 0
cell 1 info dict, finds the first key in that dictionary and uses the info keys
found within that cell's info dictionary.

* display brush location when not actively drawing in edit mode

Adds in 3rd image (brush_view) to be displayed in edit_mode on top of
annotated and raw images. brush_view shows where brush is hovering and upon
click and drag, shows where current brush stroke has been. brush_view resets
when mouse button is released. brush_view uses the same cmap as the annotation.
brush_view does not update upon brush_resize; the mouse must be moved slightly
for the updated brush to display. brush_view may also lag slightly behind mouse
movement if user is moving mouse quickly.

* add helper functions for cell info manipulations

makes code shorter and easier to read

* use helper functions to add or delete cell info from dict

Each function (in npz mode) that adds or removes a cell from a frame
or whole npz now uses helper functions to do so. This streamlines
the code and reduces chances of one add/del piece working incorrectly.

Also, making a new label from keybind commands (eg, "new cell stack")
is now based on the max label in the npz and not num_cells, as using
edit_mode to add in annotations may result in mislabeling based on num_cells.

* change vmax of annotations to correspond to highest cell value

If a new annotation is added with a higher value, the cmap scaling
will be adjusted based on that value, not on the number of cells in
the npz. Not a major change to display behavior since most of the time
the new annotation will be num_cells + 1

* add save_as_trk option to npz mode

Save_as_trk will create a lineage and save the current channel
and feature together as a .trk file. The filename will be the name
of the npz, plus "_c{}_f{}" to indicate which channel and feature
were used. The lineage file will have frame and label information
but no parent/daughter information.

Option to save as .trk will only be displayed if an npz is open.
mode.py will be passed the argument "filetype = npz" to toggle
the change in display behavior.

* add hole_fill to trk mode

Select a cell and click "F" to prompt hole fill. The next
click onto background will fill all connected pixels of
background with selected label. Lineage does not update.

* move highlight toggle out of mode_handle into on_key_press

Pressing "h" to toggle highlight mode on and off does not
use mode.py to display text or use self.mode.kind to determine action,
so it should not be under mode_handle.

* use helper functions to add or delete cells from lineage in some cases

Functions that add or delete a cell can call helper functions for
straightforward addition or deletion of cells in a lineage. This works well
for functions like action_delete and action_watershed. Functions with more
complicated lineage manipulations, like action_replace or action_new_track
will retain their own specific lineage-editing code. Helper functions will
make other functionality easier to add in, particularly edit_mode.

* create new track in just one frame

It is sometimes necessary to relabel a cell in just one frame
of a trk file. Adds in option to create single cell track by
pressing "C" (with one cell selected), then "S" (for "single")
instead of "SPACE" to confirm.

* prevent duplicate daughters when assigning parent

Uses numpy.unique to prevent adding duplicates to list
of daughters when assigning parent/daughter relationship.

* fix formatting in mode.py prompt

* update brightness scrolling in trk and npz modes

Scales brightness adjustment based on self.max_intensity
instead of a fixed value (some raw data may not be normalized).

* add edit mode (pixel labeling with brush) to trk editor

Port of npz editor's edit mode to make compatible with trk files.
Does not allow brush value to go higher than num_tracks + 1.

* add "pick color" tool to edit mode (trk and npz editors)

While in edit mode, use "p" (for "pick") to switch to a
mode prompt. Then click on a non-zero value to change the edit_value
(the value the brush can draw or erase) to the selected value. Should
make annotation fixes faster by removing need to scroll through values,
especially in files with many annotations.

* set minimum self.scale_factor to 1 (trk and npz)

Prevents divide by 0 error that crashes Caliban if self.scale_factor = 0
when self.on_mouse_motion is called. (This can happen when any file is
being opened.)

* clarify what replacing a cell does (trk and npz)

* clear mode after selecting "save as trk" (npz mode)

* remove extraneous lines when setting self.max_intensity (npz)

* add option to sequentially relabel cells in frame of npz

Sequentially relabels cells in the current frame and feature
of npz, starting from 1. Can be useful when correcting sub-images
taken from larger image set or movie, when annotations don't
necessarily start from 1. Not added to trk mode because this only changes
annotations in one frame, which would mess with sequential frames in
a trk.

* update predict_zstacks to add only non-zero values

Slight change to make first relabeling step more clear. In first
pass of matching cells between frames, relabeled_image would
get "updated" with zeros if the background was the closest match
to a cell. Now, relabeled_image is only updated with non-zero values
and the step is explicitly skipped when the closest match is background.

* update README

-Add that npz files can be saved as 'X' and 'y',
not just 'raw' and 'annotated'
-Add "relabel" to "Edit Operations" section
-Add "color picker" to "Annotation Mode" section
-Update "save" section with info about saving npz as trk