add invert option to edit mode (npz)
Hit "i" to toggle whether the raw image should be inverted.
Image is inverted by default to have dark fluorescent regions
against a light background, but dim fluorescence can be hard to see.
Turning invert off can make dim fluorescence easier to see, but
makes it more difficult to see when annotations are against
the background (and not fluorescence).