add thresholding-based prediction to edit mode (npz)
Images with decent contrast between foreground and background can
be segmented more easily with simple thresholding. With this edit-mode
tool, user selects "t" to prompt steps in threshold prediction. User
then draws a bounding box (click and drag, release to finalize box) to
select area to threshold. That area of the raw image gets thresholded and
an annotation with unique label is created. The predicted label will not
draw over existing labels. The predicted label gets added to the annotation
without user confirmation but can be easily removed since it has a unique label.
Thresholding is intended for quick annotation of regions with a clear difference
between foreground/background and likely will not be useful for separating
clumps of cells.

Note: thresholding strategy and algorithm(s) used may not be appropriate
for all datasets--further adjustment may be required depending on the image
quality.