change self.show_bounding_box to self.show_brush for edit mode (npz)
self.show_bounding_box (bool) was used to determine whether or not
a brush preview or bounding box preview should be shown to the user
upon mouse use in edit mode. More features will be added to edit mode
that utilize predictions and previews, so instead of tracking each
one of those, the preview logic is simplified by checking for
self.show_brush in combination with self.mode.action.