add conversion brush to edit mode (npz)
Conversion brush is a mode of brush use where instead of
drawing against the background (or erasing a cell), one
value gets overwritten by another. Specifically intended
for cases where the area of interest should be annotated
but the label is wrong (ie, the edge of a threshold prediction).
In edit mode, "R" toggles the color picker for the conversion brush.
First select the color that will be overwritten, then the color
that will be used to overwrite. (In this order to avoid accidentally
overwriting a patch of annotation upon color selection.)
Conversion brush will not draw against background or change
annotations aside from the target. There is no eraser mode for
the conversion brush; to switch target and brush value,
conversion mode must be triggered again. After drawing in
conversion mode use ESC to return to normal brush behavior.