speed up edit mode by selectively updating composite (npz)
Edit mode lagged a lot, likely because the composite image displayed
in edit mode was calculated in draw_current_frame(), which is called
in response to every window event (mouse movement, key press, etc).
An attribute was added to store the last-calculated composite image,
and composite calculation was moved outside of draw_current_frame to
a helper function, helper_update_composite(). The helper function is
called to update self.composite_view after actions that could affect
the composite (frame change, filters, brightness adjustment, etc),
except for on_mouse_drag as this caused a lot of lag.

helper_update_composite uses another helper function to actually
make the composite, in the hopes that helper_make_composite_img
could be used in the future to update only portions of the image
as needed.