check for mode kind before checking for mode action (npz)
Several actions check for self.mode.action to determine what behavior
to use; this can crash Caliban if the mode has been reset to None.
To prevent this, self.mode.kind is checked before looking for the
action attribute. Also, ESC in edit mode sets self.show_brush back to
True (the typical behavior when not in the middle of different Mode
actions): this should return display behavior to normal after exiting
an action partway in edit mode, as well as reducing the chance of other
Mode errors.