change raw vmax minimum value to always be above min in frame (npz)
Previously the max_intensity for each channel had a minimum of 2,
but this could cause Caliban to crash if the max_intensity were
adjusted down past whatever the minimum value of the frame was.
on_mouse_scroll should now always keep the vmax of the displayed
image above vmin, preventing the display error from crashing Caliban.