add flood contiguous cell function as watershed alternative (npz)
Ctrl+click on a cell label (with nothing selected) to bring up the
option to flood fill it with a new label. If confirmed with space,
contiguous regions of that label will be replaced with a unique
cell label in that frame. Alternative to watershed for separating
2+ cells that have duplicate labels in a frame. Connectivity param
is different from the hole filling function so that barely-connected
pieces of cell will be flooded as well; flooding cannot be used
to separate two cells that share even one diagonal pixel.