update threshold cell creation behavior to match other functions (npz)
Use max label of cell_ids + 1 to create new label when using threshold
to create cell label instead of instead of num_cells. Using max label + 1
matches behavior of other cell-creating functions and keeps the new label
a reasonable value, since expected use of this function is to replace or
delete the new label most times it is generated.