shift-click on a cell to trim stray pixels from it (npz)
Use shift-click to quickly get rid of stray pixels of selected label.
Shift-click on the cell you want to keep and space to confirm trim.
Any pixels of that label within frame that are not contiguous with
the label you clicked on will be deleted from frame. Alternative
to creating contiguous cell with flood cell and deleting remaining
stray pixels. Intended to quickly clean up stray pixels that may
result from thresholding.