flood_contiguous removes cell from info dict if necessary (npz)
If using ctrl+click to flood an annotation removes the old label
from the frame, the old label info will be updated with del_cell_info.
The behavior should match creating a cell in a single frame. Only
applicable if the contiguous cell selected is also the only instance
of that label in the current frame.