highlight shift- or ctrl-clicked cells (npz)
If highlight mode is turned on and a cell is shift- or
ctrl-clicked to modify the annotations, the cell will now be
highlighted. Useful to see which pieces might be deleted with
shift-click, or if it is necessary to use ctrl-click to split
a contiguous piece off of the label.