use array instead of color+colormap info to show label in browser (npz)
Pass the annotation values of frame (array of arrays) along with
images of frame when fetch_and_render_frame() is called. Then the
javascript portion of code can figure out which label is being
moused over by using the mouse coordinates as indices of the array.
Removes need to generate and keep track of colormaps in ZStackReview
class, and can use matplotlib colormaps for annotations. Using
color of annotations could cause problems for cells that happen to have
same color and makes handling highlighted cells trickier.