use screen_scale attribute for y and x conversions (npz)
Screen scaling was hardcoded into the js as 2, but this should
be flexible in case the screen scaling changes from file to file.
Send screen_scale factor to javascript upon opening file (expect
that screen_scale will not change once file is opened in browser)
and use that instead of a hardcoded value.