update handle_draw to reduce lag in browser (npz)
Drawing in edit mode is one of the worst times for a user to
experience lag. To reduce slowdown, changed handle_draw to only
communicate with server upon mouse release. During click&drag
while editing, a list of coordinates of mouse locations is added to.
The list of coordinates (and other info) is sent to the python
method action_handle_draw, which applies the changes and updates
cell info if a label has been added or deleted. (action_handle_draw
also now uses helper functions to add and delete cell info as needed.)