show brush preview in browser edit mode (npz)
Use brush object with draw method to draw a circle onto
a hidden canvas when in edit mode. Hidden canvas can then be
overlaid on top of the edit mode canvas with adjustable compositing
or alpha settings. (Could probably do this with the single canvas,
but a hidden canvas already exists, and this way it's straightforward
that hidden canvas = previews in edit mode.) Brush does not clear
during click&drag so that the path of the brush remains on screen.
Brush view clears when mouse button is released.

Brush color and opacity currently hardcoded in. May want to
change so that opacity is user-adjustable (slider on screen?)
and color updates to match color of cell that is being annotated.