update bucket/path setup
Previously, bucket names were hard-coded in and the structure did not allow users to be able to define folders within buckets. This update allows users to define input bucket, output
bucket, folder paths, and filename in the url in the following form:

caliban.deepcell.org/input-bucket__output-bucket__folder-path-1__folder-path-2__ ... __filename

where the first string is the input bucket, the second string is the output bucket, and the last string is the filename. Strings can not be separated by '/' because the url router treats each string in between as a separate variable, and sets of files won't always fit in the same path format. The solution to this was to have strings to be separated by '__' to account for folder/file names that have '_'.