reduce server calls by storing edit_value in js, not caliban.py (npz)
The edit_value variable (sets the value that brush will draw or erase
in pixel-editing mode) had been stored in the python object created
by caliban.py, so updating this whenever the user changed the value
with -/= required a call to the server. Storing this variable in the
javascript code removes the need to call the server when it is changed.
render_log() is called after edit_value changes so display updates
immediately with correct edit_value.