reduce server calls by storing eraser state in js (npz)
Erase variable can be stored in js and passed to python object
when it is needed, instead of storing it as part of python object
(which requires making a call to the server when the value is changed).
Removed erase attribute and action_change_erase code from ZStackReview
class since they are no longer needed to update erase variable.