fix folder path for working with test files (npz)
The landing page for Caliban was not updated to reflect the new way of accessing files and had trouble uploading the files. The dropdown list of files on the landing page are now all preloaded into a subfolder called 'test' in the input S3 bucket. To add more files to the landing page, developers only need to update the dropdown menu to add the file option to the list and upload the file to the 'test' subfolder. This commit allows all files on the landing page to be accessible.