add another npz to test file selection
Added a cytoplasmic HeLa movie npz to landing page for browser Caliban.
This npz is from the dataset that will be launched to fig8 contributors
so it is useful to make sure that it is displayed correctly/all features
work on this file before launching job. Also added some description,
in comments as well as displayed text, to make it easier to pick an
appropriate file to load for testing/demos.