add keybind to edit mode that sets brush to unused value (npz)
Adds a keybind (n) to edit mode so that annotators can easily
set the brush to be an unused value. To do this, a map (dictionary)
is added to the javascript that keeps track of the max value in
each feature whenever new "tracks" (cell info dicts) are received
from the server. The keybind changes the value of the brush to
be the max value +1 for that feature, ensuring a unique number even
if labels have been skipped.