change colormap to viridis (npz)
It can be difficult to distinguish the dark values of the cubehelix
cmap from background. Viridis is a perceptually uniform sequential
cmap that ranges from purple to blue to green to yellow (no red).
In viridis, 0 is not black; to create a clearer distinction between
foreground and background, background is masked with black in caliban.py.
(Since highlighting is done in javascript, masking in caliban.py does
not conflict with this.) This cmap should also make edit mode slightly
easier to see. Since red is not included in this cmap, no changes
to the default highlight color in js are needed.