move shift- and alt-click out of key handling and into click handling (npz)
Previously, canvas event handling would check for the shift or alt keys
upon mousedown (beginning of click) and call functions in key handling
to deal with those events. Added evt as an arg for mode.click() and
moved the shift- and alt-click functionality within mode.click(). This does
not affect other parts of mode.click(), which don't need evt info.

This groups the "click" functionality together better and lets us take advantage
of the existing logic to determine selection. Before, shift- and alt-click
would be called regardless of current mode, which could lead to buggy behavior.
Now, shift- and alt-click are only called if nothing else was selected upon click.
Additionally, shift- and alt-click now highlight the clicked cell (if highlight is on).