update zstack prediction for better prediction (desktop)
Previous version of zstack prediction struggled with multiple
overlapping cells and had buggy behavior for some relabeling of
unmatched cells. Changes include:

- if more than one cell in next_img matches the same cell in img,
the pair of cells that have the highest iou will be paired. Previously,
the first cell in next_img that matched a cell in img would get relabeled
to that cell in img. This led to some poor zstack matching when multiple
cells overlapped, in which case the pair of cells chosen depended arbitrarily
on their labels and not their iou compared to other cell pairs.

- next_img gets sequentially relabeled before any prediction starts. Doesn't
change outcome (especially because of above change), reduces iou calculations
that are done to compare the two frames, and keeps labels from being skipped
in output relabeled_img

- pared down overly complicated system for determining how to relabel cells
from next_img that had no matches (eg, are probably new cells)

- added threshold arg (default = 0.1). Cells may have small regions of overlap
that cause iou > 0, but very small overlaps are usually an accident and not
useful for determining matches between pairs of cells. Cell matches where iou
falls below threshold are considered not to have matched anything.