add KeyStateHandler to window (npz)
Not used for anything yet but can be used to modify other events
in other ways (eg, hold down space while dragging with mouse to pan,
or hold down a modifier key during mouse scroll to adjust brightness
or zoom, etc). Events that don't have access to modifiers, or should
be modified by keys not included in modifiers, can query the key_state
dict to access the currently pressed keys.