create CalibanBrush class
Holds brush values, status of eraser, and brush size. Methods are used
to update the attributes as needed. (Some methods also take another object
as an argument, eg so that brush is not increased larger than the current
dimensions of window, or to set an upper bound on edit_val.)