move on_mouse_motion method to CalibanWindow class
Mouse motion does two things: updates the position of the mouse, and redraws
brush preview if appropriate (eg, not waiting for user to draw bounding box for
thresholding). Since update_mouse_position helper method belongs to parent class
(CalibanWindow), and CalibanBrush methods are used to update the brush view,
this event is abstracted enough that it can be moved to the parent class.