move color-picking helper methods into CalibanWindow (parent) class
Methods from ZStackReview to pick color and set conversion brush values
are abstract enough to be called by either TrackReview or ZStackReview and
have been moved into CalibanWindow. TrackReview does not yet have way to
access conversion brush functionality (no keybind to start process) but if
this keybind is added, it will be trivial to make the conversion brush work with
trk files.