move draw_cell_info_label method to CalibanWindow (parent) class
Uses a mix of general and child-class specific methods to generate a pyglet Label
containing info about the current label being viewed, and draws the pyglet Label
in the bottom left of the screen (in the blank sidebar space). Moved to parent class
as all child classes are expected to need label-specific info display.