troubleshooting docker-compose, order of containers matters
eb_application tries to communicate with server as it starts up, server
must exist for that to work. This doesn't completely fix problem as server
is not ready to go as soon as image is built/run; must still use docker-compose
restart app once MySQL docker image is fully set up.