first connection to MySQL server makes sure database exists
eb_application.py does not assume that projects table in database exists,
but it does assume that database in server exists, which causes everything
to fail when working with a new Dockerized server. Fix by connecting for the
first time without a specified db and attempting to create db if it doesn't
exist on server yet.