Add outline capability to browser (#122)
* send outline info from server to frontend as negative labels

Modify the label array sent to frontend so that edges of labels have
the negative value of the label. Allows them to be grouped together
with label interiors (absolute value is still label), to isolate the
edges of one particular label, or to view all outlines at once (values
less than zero). Doesn't increase size of payload. Send array as int16 dtype.

* example of how to apply outlines on top of image