dilate watershed result if it only produces a few pixels
If watershed splits label and produces a label (original or new label)
with fewer than 5 pixels, dilate the small region to make it easier to
select. 5 pixels is somewhat arbitrary; small enough to be difficult to
select to replace or delete, doesn't dilate to be too large. Uses disk
dilation with radius 3. Intended as ease of use failsafe for when watershed
performs poorly.