modify scaled image recoloring helper to accommodate translucency
Underlying dtype for image data is clamped 8bit array, so values will
never overflow. Helper function recolorScaled takes advantage of this by adding
r, g, and b values to image data instead of setting image data to those values.
Still allows us to set values to 0 or 255 by adding either -255 or 255, but also
allows adding "translucency" by adding a fixed value (in this case, 60) to the
R, G, and B values of the pixel to modify.