Fix npz upload issues (#139)
* save npz from file-like object, not an actual file

* add health check in flask for eb deployment to use

Add health check route to flask that returns a success with 200 status code.
Configure eb deployments to use that route for health checks, which prevents
server instances from being used before they are ready.

* requested changes

* write trk tarfile to memory instead of disk for uploading step

Saving the file on disk is not intended behavior, we just need to
send the updated file object to the output s3 bucket.

* remove unnecessary import (secure_filename)

secure_filename creates a safe path for saving files to disk to prevent
vulnerabilities. This app no longer saves files to disk, creating temporary
file objects instead, so we don't need this.