Cursor, viewing window js classes (#196)
* initial commit of zoom/pan handler and cursor classes

Trying to group mouse/cursor attributes and methods together. However,
mouse location within image depends on current zoom and offset of
displayed image, so making zoom and pan methods more modular will likely
make mouse attributes and methods cleaner as well.

* include cursor.js when loading page

* replace mousedown global var with cursor attribute

* add viewer object to zstack js

* switch to using viewer sx, sy attributes; update with viewer.pan

* use viewer.zoomLimit instead of global var

* add trace attribute to cursor, replace global var

Doesn't change how trace is handled, just a scope change for now.

* add width, height, scale attributes to cursor object

Will be needed for calculations of position in image from position on
canvas. Update scale in window resize event if needed.

* move inRange checking function into cursor class

* store scale as a brush attribute, update on image rescale

Brush is currently constructed with scale as one of the inputs but
it isn't used for anything. It will be used for magnification calculation
in brush.draw implementation change so it should be updated appropriately.

* add notes to updateMousePos

* use viewer zoom attribute instead of global var

* add scale to viewer object, use instead of global var

- construct object in load_file after dimensions are received but
  before scale is set; construct with placeholder scale=1
- update viewer.scale in setCanvasDimensions

* remove scale attribute from brush object

scale will be passed in via viewer object if needed

* pass viewer object to brush.draw instead of the separate attributes

simplifies brush.draw command

* move canvas border attributes, setting, and drawing to viewer object

* move most of changeZoom functionality into viewer method

Needs canvasPos coords to be passed to it

* move _rawMouse coords to cursor object and rename

* move canvasPos coords to cursor object

* remove args from cursor.inRange

since canvasPos coords are now an attribute of the object,
nothing needs to be passed in to check whether the cursor is
hovering over a displayed part of the canvas or not

* move imgX and imgY to cursor instead of global vars; change updateMousePos

updateMousePos now calls cursor.updatePos to update cursor location
attributes, then updates other objects as needed if cursor location
has actually changed

* remove label_under_mouse in favor of cursor method

segmentation array is passed to cursor.getCurrentLabel and the current
label is returned; this gets whenever render_info_display is called,
which is also called whenever render_image_display is called. a possible
alternative would be moving this functionality further into the cursor
class so that cursor.label is an attribute updated on cursor position change

* add storedClick coords to cursor instead of using global var

* add comments/TODOs

* add segArray as cursor attribute, update when segArray changes

allows cursor to store the label it is currently hovering over as
one of its attributes

* replace current_label global var with cursor.label, add segArray to cursor

- doesn't get rid of seg_array as global var
- cursor gets segArray attribute so new label is calculated when
  cursor location changes (in cursor.updatePos)
- no more need for cursor.getCurrentLabel()
- getter and setter for segArray so label updates upon changed frame

* Merge Cursor and View objects. (#198)

* add ESLint dependencies

* merge the cursor and the view into a single canvas state object.

They use a lot of the same variables and are interdependent on eachother for shared state. Might as well make a single State object.

* Refactor load_file to loadFile and pass it a callback inside startCaliban.

* Finish up the viewer changes.

* use vanilla javascript selectors where possible.

* ESLint global functions

* ESLint Mode class.

* rename cursor.js to canvas.js

* fix various bugs with viewer/cursor

* a bit of variable linting

* add helper functions for calcMaxHeight and calcMaxWidth.

* remove prepare_canvas function and add event listeners directly in the loadFile callback.

* fix rendering bug

* pruning jquery selectors where possible, fixed a couple event handler bugs.

* remove dimensions and seg_array from global variables, using viewer attributes instead.

* move onLoad definitions from adjust.js to startCaliban to remove globals from adjust.js

* ESLinting Mode some more.

* call clearTrace!

* make initial loadFile asyncrhonous.

* refactor CalibanCanvas to CanvasState and viewer to state.

I think its a more clear name, open to feedback.

* final ESLinting.

* Fix RGB bug.

* set to true/false directly to avoid any confusing state situations.

* fix handleMousemove event handler

* tracks is used in render_cell_info

* fix segArray bug.

* pan() should not always scale the dx/dy inputs.

* use the new scaledWidth property more.

* fix spacedown toggle and remove unused toggle methods.

* remove toggle helper methods that are no longer used.

* remove redundant method.

Co-authored-by: willgraf <7930703+willgraf@users.noreply.github.com>